requirements.txt
Django
djangorestframework
django-compressor